make few commonly used components 'common' so they can be used across the whole app

+++ src/components/common/DeleteForm.js
@@ -2,7 +2,8 @@ import React from 'react';
 import {
   useToasts
 } from 'react-toast-notifications';
-import Api from '../api/Api';
+import Api from '../../api/Api';
+import AdminApi from '../../api/AdminApi';
 import './DeleteForm.css';
 
 function DeleteForm(props) {
@@ -16,24 +17,27 @@ function DeleteForm(props) {
     }
 
     // parse input
-    let selected_ids = props.selected.map(post => post.id);
+    let selected_ids = props.selected.map(item => item.id);
 
     try {
       // send delete requests
       for (let id of selected_ids) {
-        await Api.deletePost(id);
+        if (props.type === 'post')
+          await Api.deletePost(id);
+        // else if (props.type === 'admin_post') ...
+        // else if (props.type === 'admin_report') ...
       }
 
       // refresh page
       window.location.reload();
     } catch (err) {
-      addToast('Failure deleting post, status: ' + err.status, {appearance: 'error'});
+      addToast('Failure deleting item, status: ' + err.status, {appearance: 'error'});
     }
   };
 
   return (
     <form className="delete_form" onSubmit={onSubmit} autoComplete="off">
-      <label>Delete Post</label>
+      <label>{props.text}</label>
       <input id="submit_delete" type="submit" value="Submit"></input>
     </form>
   );
